test(cli): align binary view tests with current conventions
Group the newly added binary view coverage in BDD suites and use expect
matchers so failures show structural differences. Rename the load-input
test to describe behavior instead of its coverage purpose, and preserve
conditional registration for Linux procfs cases.
