+++ packages/cli/test/view-procfs.test.ts
@@ -1,15 +1,17 @@
-import { assert, assertEquals } from "@std/assert";
+import { assert } from "@std/assert";
+import { expect } from "@std/expect";
+import { describe, it } from "@std/testing/bdd";
 import { join } from "@std/path";
 import { realFileGateway } from "../lib/view/filegateway.ts";
 import { loadViewInput } from "../lib/view/loadinput.ts";
 
 const PROCFS_PATH = "/proc/self/cmdline";
 
-Deno.test({
-  name: "binary input: zero-sized virtual files are read through EOF",
-  ignore: Deno.build.os !== "linux",
-  async fn() {
-    assertEquals(Deno.statSync(PROCFS_PATH).size, 0);
+describe("view-procfs", () => {
+  const linuxIt = Deno.build.os === "linux" ? it : it.skip;
+
+  linuxIt("reads zero-sized virtual files through EOF", async () => {
+    expect(Deno.statSync(PROCFS_PATH).size).toBe(0);
     const input = await loadViewInput(
       PROCFS_PATH,
       PROCFS_PATH,
@@ -18,43 +20,33 @@ Deno.test({
       false,
     );
 
-    assertEquals(input.kind, "bytes");
-    if (input.kind !== "bytes") return;
+    assert(input.kind === "bytes");
     assert(input.bytes.length > 0);
-    assertEquals(input.language?.id, "binary");
-    assertEquals(input.extent, {
+    expect(input.language?.id).toBe("binary");
+    expect(input.extent).toEqual({
       byteLength: input.bytes.length,
       complete: true,
     });
-  },
-});
+  });
 
-Deno.test({
-  name: "realFileGateway.open: reads a zero-sized procfs file",
-  ignore: Deno.build.os !== "linux",
-  fn() {
-    assertEquals(Deno.statSync(PROCFS_PATH).size, 0);
+  linuxIt("opens a zero-sized procfs file through the real gateway", () => {
+    expect(Deno.statSync(PROCFS_PATH).size).toBe(0);
     const opened = realFileGateway().open(PROCFS_PATH);
     assert(opened !== null);
     assert(opened.text.length > 0);
-    assertEquals(opened.source.editable, false);
+    expect(opened.source.editable).toBe(false);
     const rendered = opened.source.render?.(
       opened.source.parse(opened.text),
     );
     assert(rendered !== undefined);
     assert(rendered.lines[0].text.startsWith("00000000  "));
-    assertEquals(
-      rendered.lines.at(-1)?.text.includes("preview stopped"),
+    expect(rendered.lines.at(-1)?.text.includes("preview stopped")).toBe(
       false,
     );
-  },
-});
+  });
 
-Deno.test({
-  name: "cf view reads a zero-sized procfs file",
-  ignore: Deno.build.os !== "linux",
-  async fn() {
-    assertEquals(Deno.statSync(PROCFS_PATH).size, 0);
+  linuxIt("renders a zero-sized procfs file through `cf view`", async () => {
+    expect(Deno.statSync(PROCFS_PATH).size).toBe(0);
     const result = await new Deno.Command(Deno.execPath(), {
       cwd: join(import.meta.dirname!, ".."),
       args: [
@@ -72,9 +64,9 @@ Deno.test({
     }).output();
 
     const decoder = new TextDecoder();
-    assertEquals(result.code, 0, decoder.decode(result.stderr));
+    expect(result.code, decoder.decode(result.stderr)).toBe(0);
     const stdout = decoder.decode(result.stdout).split("\n").filter(Boolean);
     assert(stdout[0].startsWith("00000000  "), stdout.join("\n"));
-    assertEquals(stdout.at(-1) === "00000000", false);
-  },
+    expect(stdout.at(-1) === "00000000").toBe(false);
+  });
 });
